+++ src/screen_manager.h
@@ -13,7 +13,7 @@ class ScreenManager
   public:
     void show_screen(std::shared_ptr<IScreen> screen);
     void close_screen();
-    void update(float delta_time);
+    void update(double delta_time);
     int stack_size();
     void draw(Renderer &renderer, bool lightweight = false);
